Wrap clipboard functionality in a try-catch block
Moved the clipboard copy operation inside a try-catch block to handle cases where the operation might fail or an exception is thrown. Before this change, the lack of proper error handling could have resulted in a less than user-friendly experience. Now, the script will try to execute the copy operation, and if unsuccessful, it will throw an error message. In the case of a successful operation, it provides visual feedback by changing the link color to green.