fix(theme): change dark theme functionality
We had a bug where users would see a white theme flash while using dark
theme. It was happening because our server didn't know which theme users
want to use. Therefore, JavaScript was responsible for theme change.
But, since JavasScript is not loaded immediatly but after some delay, we
whould see a white theme for a half of a second.

Now, our server knows the theme user prefers, it sets a CSS class to
<html> tag and CSS can set the correct theme